QA final project/pyproject.toml
[project]
name = "storeappeldar"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = ["mysql-connector-python", "flask", "selenium", "requests"]
